Move add metadata to constructed after ctf initialization
This fixes crash of non-periodic tracers that were adding the event
metadata at tracer init where ctf is not initialized yet, now this
moves that logic to the tracer constructed after the parent constructed
ensuring that ctf is initialized before the metadata is added. The tracers
modified are: interlatency, proctime, queuelevel and scheduletime.